fix: Set default owner on safe creation when entering step
